Add to admin ability to create new cars on map

+++ logic/src/main/java/com/olrox/car/domain/Coordinates.java
@@ -8,9 +8,9 @@ public class Coordinates {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
 
-    private int latitude;
+    private double latitude;
 
-    private int longitude;
+    private double longitude;
 
     @OneToOne(mappedBy = "coordinates")
     private Car car;
@@ -23,19 +23,19 @@ public class Coordinates {
         this.id = id;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
